test: 테스트 메서드 위치 수정 & 테스트 추가
